Implemented SearchText position in redux store
New folder was created in store - filters. There wil be defined all
action screators and reducers which apply to the way of displaying
evnets on the page.
File actionTypes.js was created to define types of actions (with one new
position: filters.SEARCH_TEXT_CHANGED.
File actions.js was created with one action (searchTextChanged) which
will be used when user is typing a search phrase.
File reducer.js was created - it manages to change redux store on a
basis of filter actions. It also has selectors which allow components to
take information from store as props (getCurrentSearchText selector was
defined).