Created selector getEventsFetched
Selector takes eventsFetched object from the store and converts it to
the array. Array consists of events as object with added id as a
previous key value.
In next step array is sorted by the selected sort method from the store.
At the end array is filtered by the search prhase from the store.
Sorted and filtered array is returned by the function.