Added action creators: eventUpdated, eventDeleted
Action creators were added in actions.js files. eventUpdated requiers
eventData and eventId parameter. EventData is an event as object defined
in exact the same way as on firebase server. EventDeleted requires id of
the event to remove.
Actions are handled in reducer.js file. In both situations initialState
is returned - eventForm is cleared and editingMode is disabled.