Added function to edit event in SingleEvent.js
SingleEvent has been connected to new action creator, which is
dispatched when user presses button to eddid the event in
EventDaataButtons component. Action is dispatched, which sets up editing
form and puts it in the state of eventEditing. Clicking on button
redirects user to AddEventForm component, which is uded to edit the
event.

All other operations are made in AddEventForm component.