Added function to handle event editing in AddEvent component
Compoent AddEvent could be used for event editing - new data was added
to the store (edited Event ID, editing finished). It was passed to the
AddEventForm through props, which changes it's look depending on the
values.
New action 'startEventEditing' which is dispathced form SingleEvent
site, fills up the form with specific event data. Action
'sendEditedEvent' is dispatched when user finish edit the event.
After successfull event updating popup is shown and user is redirected
to the event site.