Evnet distance from the user added
Every event in BrowseEvents components has distance calculated basing on
geolocation service.

New action creator was added in events domain - getEventDistance. New
position was added to the store: eventDistances - which is an empty
object. When this action is dipstached, distance is calculated and
reducer adds it to the evenDistances as a key:value par, where key is
eventId and value is the calculated distance.

Action is hooked to the BrowseEvents and passed to EventTab cmoponent.
This component checks if user enabled his position and if the distance
was calculated already - if not it dispatches an action and displays
distance when calculated.