Added all events map
Container component EventMap.js displays map with every defined event on
it. It fetches data of all evnets when after mounting, and passes it to
new component - AllEventMap. This component processes data and creates
markes on the map with information about every event. Popups from
markers have links to the events.
AllEventMap is wrapped by AllEventMapWrapper component - it's necessary
for googleMaps Api to work correctly.